+++ src/test/java/com/jstarcraft/ai/neuralnetwork/learn/IgnoreLearnerTestCase.java
@@ -10,14 +10,14 @@ import com.jstarcraft.ai.neuralnetwork.learn.Learner;
 public class IgnoreLearnerTestCase extends LearnerTestCase {
 
 	@Override
-	protected GradientUpdater<?> getOldFunction(int[] shape) {
+	protected GradientUpdater<?> getOldFunction(long[] shape) {
 		NoOp configuration = new NoOp();
 		GradientUpdater<?> oldFunction = new NoOpUpdater(configuration);
 		return oldFunction;
 	}
 
 	@Override
-	protected Learner getNewFunction(int[] shape) {
+	protected Learner getNewFunction(long[] shape) {
 		Learner newFuction = new IgnoreLearner();
 		return newFuction;
 	}
