+++ src/test/java/com/jstarcraft/ai/neuralnetwork/learn/NadamLearnerTestCase.java
@@ -12,7 +12,7 @@ import com.jstarcraft.ai.neuralnetwork.learn.NadamLearner;
 public class NadamLearnerTestCase extends LearnerTestCase {
 
 	@Override
-	protected GradientUpdater<?> getOldFunction(int[] shape) {
+	protected GradientUpdater<?> getOldFunction(long[] shape) {
 		Nadam configuration = new Nadam();
 		GradientUpdater<?> oldFunction = new NadamUpdater(configuration);
 		int length = (int) (shape[0] * configuration.stateSize(shape[1]));
@@ -22,7 +22,7 @@ public class NadamLearnerTestCase extends LearnerTestCase {
 	}
 
 	@Override
-	protected Learner getNewFunction(int[] shape) {
+	protected Learner getNewFunction(long[] shape) {
 		Learner newFuction = new NadamLearner();
 		return newFuction;
 	}
